[Refactor] useList 구조 수정
